requirements.txt
httpx
feedparser
pydantic
python-dateutil
bleach

# dev
yapf
pytest
ruff
black
